Bug fix build errors (#65)
The Gulp build process failed because the last Dependabot update changed
the package in use. This caused us to have to install a couple of other
packages, that we previously didn't need to install. And by having to
install them, we had to rewrite our logic for that step.

We were getting errors before, but now it's working as expected. 

<img width="600" alt="image"
src="https://github.com/user-attachments/assets/6c740305-2c45-47e5-83ce-918f8b28dc2b"
/>